refactor: remove reth_stateless conversion and retain necessary for zkboost
