+++ Cargo.toml
@@ -40,8 +40,8 @@ result_large_err = "allow"
 [workspace.dependencies]
 anyhow = { version = "1.0", default-features = false }
 clap = { version = "4.6.0", features = ["derive"] }
-hex-literal = { version = "1.1.0", default-features = false }
 const-hex = { version = "1.19", default-features = false }
+hex-literal = { version = "1.1.0", default-features = false }
 once_cell = { version = "1.21", default-features = false }
 serde = { version = "1.0", default-features = false }
 sha2 = { version = "0.10.9", default-features = false }
@@ -65,15 +65,11 @@ zstd = "0.13.3"
 # alloy
 alloy-consensus = { version = "2.0.5", default-features = false }
 alloy-eips = { version = "2.0.5", default-features = false }
-alloy-genesis = { version = "2.0.5", default-features = false, features = [
-    "serde-bincode-compat",
-] }
-alloy-rpc-types-engine = { version = "2.0.5", default-features = false }
-alloy-primitives = { version = "1.6.0", default-features = false, features = [
-    "map-foldhash",
-    "map-indexmap",
-] }
+alloy-genesis = { version = "2.0.5", default-features = false, features = ["serde-bincode-compat"] }
+alloy-primitives = { version = "1.6.0", default-features = false, features = ["map-foldhash", "map-indexmap"] }
 alloy-rlp = { version = "0.3.13", default-features = false }
+alloy-rpc-types-debug = { version = "2.0.5", default-features = false }
+alloy-rpc-types-engine = { version = "2.0.5", default-features = false }
 
 # revm
 revm = { version = "40.0.3", default-features = false }
@@ -82,8 +78,8 @@ revm = { version = "40.0.3", default-features = false }
 reth-chainspec = { git = "https://github.com/paradigmxyz/reth", tag = "v2.3.0", default-features = false }
 reth-ethereum-primitives = { git = "https://github.com/paradigmxyz/reth", tag = "v2.3.0", default-features = false }
 reth-evm-ethereum = { git = "https://github.com/paradigmxyz/reth", tag = "v2.3.0", default-features = false }
-reth-primitives-traits = { version = "0.4.1", default-features = false }
 reth-payload-validator = { git = "https://github.com/paradigmxyz/reth", tag = "v2.3.0", default-features = false }
+reth-primitives-traits = { version = "0.4.1", default-features = false }
 reth-stateless = { git = "https://github.com/han0110/stateless", branch = "feature/v2.3.0", default-features = false, package = "stateless" }
 reth-tries = { git = "https://github.com/han0110/stateless", branch = "feature/v2.3.0", default-features = false, package = "tries" }
 
@@ -102,8 +98,8 @@ libssz-types = { version = "0.2.2", default-features = false, features = ["alloc
 zkvm-interface = { git = "https://github.com/eth-act/zkvm-standards", rev = "282cd356c3a0498416bb0619f9c8a347ce9933fb" }
 
 # ere
-ere-platform-core = { git = "https://github.com/eth-act/ere", rev = "2505e96eb9821b7aa561b56eeeddc145baa138fd" }
 ere-dockerized = { git = "https://github.com/eth-act/ere", rev = "2505e96eb9821b7aa561b56eeeddc145baa138fd" }
+ere-platform-core = { git = "https://github.com/eth-act/ere", rev = "2505e96eb9821b7aa561b56eeeddc145baa138fd" }
 ere-util-build = { git = "https://github.com/eth-act/ere", rev = "2505e96eb9821b7aa561b56eeeddc145baa138fd" }
 
 # local
