+++ crates/stateless-validator-test/Cargo.toml
@@ -26,14 +26,14 @@ walkdir.workspace = true
 zstd.workspace = true
 
 # alloy
-alloy-primitives.workspace = true
+alloy-primitives = { workspace = true, features = ["serde"] }
 
 # ere
 ere-dockerized.workspace = true
 ere-platform-core.workspace = true
 
 # local
-stateless-validator-common = { workspace = true, features = ["host", "legacy"] }
+stateless-validator-common = { workspace = true, features = ["host"] }
 
 [dev-dependencies]
 paste.workspace = true
